// src/api/generated/models/favoriteVO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface FavoriteVO {
  id?: string;
  userId?: string;
  name?: string;
  description?: string;
  sortOrder?: number;
  isDefault?: boolean;
  appCount?: string;
  createdAt?: string;
}